move some utils to camera-core
