+++ camera-core/src/main/java/com/stripe/android/camera/scanui/util/ViewExtensions.kt
@@ -3,8 +3,10 @@ package com.stripe.android.camera.scanui.util
 import android.content.Context
 import android.content.res.Resources.NotFoundException
 import android.graphics.Rect
+import android.graphics.drawable.Animatable
 import android.util.TypedValue
 import android.view.View
+import android.widget.ImageView
 import androidx.annotation.DimenRes
 import androidx.annotation.DrawableRes
 import androidx.annotation.RestrictTo
@@ -21,6 +23,27 @@ internal fun Context.getDrawableByRes(@DrawableRes drawableRes: Int) =
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 fun View.asRect() = Rect(left, top, right, bottom)
+
+/**
+ * Set the image of an [ImageView] using a [DrawableRes].
+ */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+fun ImageView.setDrawable(@DrawableRes drawableRes: Int) {
+    this.setImageDrawable(this.context.getDrawableByRes(drawableRes))
+}
+
+/**
+ * Set the image of an [ImageView] using a [DrawableRes] and start the animation.
+ */
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+fun ImageView.startAnimation(@DrawableRes drawableRes: Int) {
+    val d = this.context.getDrawableByRes(drawableRes)
+    setImageDrawable(d)
+    if (d is Animatable) {
+        d.start()
+    }
+}
 
 /**
  * This is copied from Resources.java for API 29 so that we can continue to support API 21.
